refactor(#185): migrate hub, errors, and manifest to canonical src seams
